Refactored get_load_balancer_name to be used when getting stack health
